与前端代码交互

+++ src/main/java/com/itjn/common/context/BaseContext.java
@@ -1,14 +1,15 @@
 package com.itjn.common.context;
 
+//线程上下文  BaseContext类使用 ThreadLocal来存储当前用户 ID
 public class BaseContext {
 
-    public static ThreadLocal<Long> threadLocal = new ThreadLocal<>();
+    public static ThreadLocal<Integer> threadLocal = new ThreadLocal<>();
 
-    public static void setCurrentId(Long id) {
+    public static void setCurrentId(Integer id) {
         threadLocal.set(id);
     }
 
-    public static Long getCurrentId() {
+    public static Integer getCurrentId() {
         return threadLocal.get();
     }
